requirements.txt
python-telegram-bot
requests
python-dotenv
fastapi
uvicorn
jinja2
python-multipart
aiofiles